[TTLG] Introduce logs management via OTEL
This commit brings the opentelemetry logs management in the Python project.

It adds the instructions to install the opentelemetry-collector component, which
is mandatory fo the OTEL scripts to work.

The new scripts are name `-otel.py` and the old one that were using the direct
writing are now named `-direct.py`.

The old files:

- python-app/backend.py
- python-app/frontend.py
- python-app/logs.py
- python-app/variables.py

are now symlink to the `-otel.py` files, so that it is fairly easy to use one or
another version to operate.
