+++ Through-The-Looking-Glass/helm-opentelemetry-collector.yml
@@ -1,0 +1,49 @@
+mode: deployment
+
+image:
+  repository: "otel/opentelemetry-collector"
+
+config:
+  receivers:
+    otlp:
+      protocols:
+        grpc:
+          endpoint: 0.0.0.0:4317
+    prometheus:
+      config:
+        scrape_configs:
+          - job_name: 'otel-collector-metrics'
+            static_configs:
+              - targets: ['prometheus-kube-prometheus-prometheus.monitoring.svc.cluster.local:9090']  # Adjust this to the appropriate target
+
+  processors:
+    batch:
+      timeout: 5s
+      send_batch_size: 1024
+
+  exporters:
+    # Tempo
+    otlp/tempo:
+      endpoint: "tempo-distributor.tempo-system.svc.cluster.local:4317"
+      tls:
+        insecure: true
+    # Loki
+    otlphttp/loki:
+      endpoint: "http://loki-gateway.loki.svc.cluster.local/otlp"
+      tls:
+        insecure: true
+
+  service:
+    pipelines:
+      traces:
+        receivers: [otlp]
+        processors: [batch]
+        exporters: [otlp/tempo]
+#      metrics:
+ #       receivers: [prometheus]
+ #       processors: [batch]
+ #       exporters: [otlp]  # Metrics are now being exported to OTLP for Prometheus to scrape
+      logs:
+        receivers: [otlp]
+        processors: [batch]
+        exporters: [otlphttp/loki]  # Logs are exported using OTLP HTTP to Loki
